drivers/qemu: Allow building Bochs display driver for non-x86 architectures
The Bochs display driver was written for x86 and is only allowed to
build only for that, but the devices can be used just fine on other
architectures on the QEMU side as long as the emulated platform supports
PCI. Commit efaf1b32ba1eb ("drivers/emulation/qemu/bochs: Rewrite
driver") adds support for using MMIO mapped Bochs display devices, which
would work for other architectures as well if not for the x86 parts.

The existing driver can use MMIO, but if the device is a legacy-capable
VGA device it defaults to using I/O ports. Add an additional CONFIG(VGA)
check to that so we can avoid x86-style I/O for architectures where that
is not implemented yet. A previous patch adds stubs for those calls to
arch/io.h headers so that this code can actually be compiled on non-x86
arches.

Allow the Bochs display driver to be built for non-x86 QEMU boards by
changing the Kconfig dependencies. Make VGA and text framebuffer support
depend on x86. Add a dependency on PCI since this is a PCI device and
vexpress-a9 (qemu-armv7) doesn't have the (emulated) hardware for PCI.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>