drivers/qemu/bochs: Allow building for non-x86 architectures
The Bochs display driver was written for x86 and is only allowed to
build only for that, but the devices can be used just fine on other
architectures on the QEMU side as long as the emulated platform supports
PCI. Commit efaf1b32ba1eb ("drivers/emulation/qemu/bochs: Rewrite
driver") adds support for using MMIO mapped Bochs display devices, which
would work for other architectures as well if not for the x86 parts.

The existing driver can use MMIO, but if the device is a legacy-capable
VGA device it defaults to using I/O ports. Add an additional CONFIG(VGA)
check to that so we can avoid x86-style I/O for architectures where that
is not implemented yet.

Allow the Bochs display driver to be built for non-x86 QEMU boards by
changing the Kconfig dependencies and using #if directives to avoid
port-based IO functions. Make VGA and text framebuffer support depend on
x86. Add a dependency on PCI since this is a PCI device and vexpress-a9
(qemu-armv7) doesn't have the (emulated) hardware for PCI.

Change-Id: I7f72d7ea13e54ecf89d067394c02b572c5f92d24
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>