mainboard/qemu-armv7-virt: Add generic virtual platform for ARMv7
The ARM Versatile Express board is actual hardware that can be emulated
by QEMU, but it lacks features like PCI, IDE, SCSI, USB support. So far
it has been the only emulated ARMv7 system in coreboot, so unfortunately
has captured the qemu-armv7 name and BOARD_EMULATION_QEMU_ARMV7 config.

Add a qemu-armv7-virt mainboard intended to be used on the QEMU ARMv7
generic virtual platform, "virt". Most of it is copied from the existing
qemu-aarch64 board, with some small parts from qemu-armv7. The resulting
build can be used with a QEMU command line like:

    qemu-system-arm \
        -display gtk,show-tabs=on \
        -bios build/coreboot.rom \
        -M virt-2.8,secure=on,virtualization=on
        -cpu cortex-a15 -m 2048M \
        -serial stdio -device VGA \
        ...

Note that the machine has to be a versioned "virt" lower than 2.11,
because the RAM discovery process in ramdetect.c relies on accessing
out-of-RAM addresses, and QEMU started generating data_abort exceptions
for these [1].

[1] https://lore.kernel.org/qemu-devel/1504626814-23124-1-git-send-email-peter.maydell@linaro.org/T/#u

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>