mainboard/qemu-riscv: Add PCI support
Add PCI support for the qemu-riscv mainboard like how it is done for
qemu-aarch64. The resulting build can be used with a QEMU command line
like:

    qemu-system-riscv64 \
        -display gtk,show-tabs=on \
        -bios build/coreboot.rom \
	-drive if=pflash,file=build/coreboot.rom,format=raw \
        -M virt -m 1G \
        -device pcie-root-port \
        -serial stdio -device bochs-display \
        ...

One complication is that the PCIe high MMIO base depends on the amount
of available RAM for 64-bit RISC-V VMs [1], but not for 32-bit ones.

[1] https://gitlab.com/qemu-project/qemu/-/blob/v8.2.2/hw/riscv/virt.c#L1474-1489

Change-Id: Ie57b35620af82c681d1f0d78fa8e514e4df0d2ac
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>