mainboard/qemu-riscv: Add PCI support
Add PCI support for the qemu-riscv mainboard like how it is done for
qemu-aarch64. The resulting build can be used with a QEMU command line
like:

    util/riscv/make-spike-elf.sh build/coreboot.{rom,elf}
    qemu-system-riscv64 \
        -display gtk,show-tabs=on \
        -bios none \
        -kernel build/coreboot.elf \
        -M virt -m 1G \
        -device pcie-root-port \
        -serial stdio -device VGA \
        ...

One current problem is that probe_ramsize() causes an error, which has
to be worked around by e.g. setting DRAM_SIZE_MB=1024 and adding a
`return 1024;` to probe_ramsize() in order for the mainboard to work.
But this is the case even before this commit.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>