mb/emulation/qemu: Move fw_cfg driver to drivers/emulation/qemu
x86 QEMU mainboards use QEMU's firmware configuration device to set up
ACPI and SMBIOS tables. The driver itself is located in the qemu-i440fx
mainboard directory presumably because it was the only one using the
device. Later on the qemu-q95 board was added, which references the
files in the qemu-i440fx directory instead of moving them to a common
scope.

The device is not specific to any architecture, though the driver is
written with x86 assumptions. We want to port it to other architectures,
so keeping the code under an x86 board is not exactly appropriate.

Move the fw_cfg driver to drivers/emulation/qemu, with a dependency on
the CPU_QEMU_X86 config option like the graphics drivers. Move its
header files into src/include and update any #include directives.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>