mainboard/qemu-riscv: Get top of memory from device-tree blob
Trying to probe RAM space to figure out top of memory causes an error on
RISC-V virtual machines with recent versions of QEMU. A previous commit
adds a library function to parse device-tree blob that QEMU provides us.
Use it to determine top of memory in RISC-V QEMU virtual machines, but
still fall back to the RAM probing approach as a last-ditch effort.

Change-Id: I9e4a95f49ad373675939329eef40d7423a4132ab
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>