drivers/qemu/fw_cfg: Support architectures using MMIO
The QEMU firmware configuration device can be used with non-x86 VMs as
well, but its driver needs to use MMIO at a platform-specific address.
On the other hand, it doesn't support MMIO access on x86 so we can't
completely switch to that even if we wanted to.

Limit the driver's port I/O function calls to x86 and implement MMIO
access as the alternative on non-x86. Require specifying the driver's
base address as a config option on non-x86 platforms, similar to how PCI
ECAM address is handled.

Change-Id: Iacab3f6959cb6af747fde8a5af5cb35ce2bca512
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>