drivers/qemu/cirrus: Add missing arch/io.h include
The Cirrus display driver uses port I/O functions without including the
arch/io.h header that defines them. The header is included through
arch/pci_ops.h and arch/pci_io_cfg.h on x86, so we don't normally get
a build error. This is not the case on other architectures.

Include the necessary arch/io.h header directly in the Cirrus driver
to help make it possible to build it on other architectures.

Change-Id: I84c76e6e723c776f45d40d98029e33adb57005c0
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>