drivers/qemu/cirrus: Allow building for non-x86 architectures
The Cirrus display driver uses port I/O functions and VGA support code
for initialization, so it could only have been built on x86 so far, but
the device can be used just fine on other architectures on the QEMU side
as long as the emulated platform supports PCI. Previous commits add port
I/O functions and enable building VGA code for more architectures
including ARM* and RISC-V, which should allow this driver to be
successfully built and used on these as well.

Allow the Cirrus display driver to be built for non-x86 QEMU boards by
changing the Kconfig dependencies. Make VGA text framebuffer support
depend on x86, because it isn't usable at the standard 0xB8000 address
on other architectrures. Add a dependency on PCI since this is a PCI
device and vexpress-a9 (qemu-armv7) doesn't have the (emulated) hardware
for PCI.

Change-Id: I168d3bf9b27d5eefa1d4844b861d73ec8437ee16
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>