drivers/pc80/vga: Allow building for non-x86 architectures
The Cirrus driver uses the VGA support code for its initialization,
which in turn depends on port I/O functions to talk to the PCI device.
Previous commits implement these and fix build errors with VGA code on
more architectures, so it's possible to have things also work there to
an extent.

However, building VGA support code is restricted to x86 architectures in
its Makefile, so trying to build the Cirrus driver for e.g. ARM64
results in linker errors. Remove the Makefile restriction.

Change-Id: Ic3199fda500c51aa3261cd9eff9a5b0b052fd742
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>