arch/riscv: Fix stack alignment to 16 bytes
On a 32-bit build for the `qemu-riscv` mainboard, trying to print values
using "%llx", "%lld", or "%llu" results in some form of corruption. For
example, if we add the following code to `mainboard_enable()`:

    printk(BIOS_DEBUG, "%d %s %lld \n", 123, "=", 123LL);
    printk(BIOS_DEBUG, "%lld %s %d \n", 123LL, "=", 123);

This prints nonsensical output and triggers an exception:

    [DEBUG]  123 = 4611686018427387904
    [DEBUG]  -9220356540316254208
    [DEBUG]  Exception:          Load access fault
    [DEBUG]  Hart ID:            0
    [DEBUG]  Previous mode:      machine
    [DEBUG]  Bad instruction pc: 0x800a6994
    [DEBUG]  Bad address:        0x0000007b
    [DEBUG]  Stored ra:          0x800a9b98
    [DEBUG]  Stored sp:          0x802a5ecc

Searching online leads to a similar issue [1] stating that the variadic
function implementation is unable to handle stack pointers not aligned
to 16 bytes, and that the stack pointer should always be aligned to 16
bytes. In the exception above, the `sp` value is indeed misaligned from
that by 12 bytes (or 4 bytes if we could go higher).

We initially align the stack pointer to the 4K page size, then reserve
space for hart-local storage at the top of it. Align the corresponding
struct to 16 bytes and update its size values to keep the stack pointer
correctly aligned.

[1] https://web.archive.org/web/20240606125616/https://github.com/riscvarchive/riscv-gcc/issues/63

Change-Id: I947ccfa2ded7e840128658d7b12297d377c340d1
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>